requirements.txt
# ============================================
# FILE: requirements_py313.txt
# ============================================
# Core AI/ML - Updated for Python 3.13
tensorflow==2.20.0  # Latest version that might work
scikit-learn==1.6.1
opencv-python==4.10.0.84
pillow==11.1.0
pydicom==3.0.1
numpy==2.2.4
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2

# Web Framework
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
flask-migrate==4.0.7
flask-cors==5.0.0
flask-mail==0.10.0
werkzeug==3.1.3
gunicorn==23.0.0

# Database
sqlalchemy==2.0.38
alembic==1.15.1

# Security
bcrypt==4.2.1
pyjwt==2.10.1
passlib==1.7.4
itsdangerous==2.2.0
pyotp==2.9.0
qrcode==7.4.2

# PDF Report Generation
reportlab==4.3.1

# Utilities
python-dotenv==1.0.1
email-validator==2.2.0
requests==2.32.3
tqdm==4.67.1
typing-extensions==4.12.2

# System Monitoring
psutil==6.1.1

# Development Tools
jupyter==1.1.1
ipython==8.32.0
pytest==8.3.5
